Issue #141 refactor: composable 함수를 모두 파일로 분리
